+++ src/CFG.js
@@ -81,6 +81,9 @@ const CFG = {};
     //    or desktop file system implementation but applied by default in
     //    browser for parity)
 
+    // File system module with `unlink` to remove deleted database files
+    'fs',
+
     // Used when setting global shims to determine whether to try to add
     //   other globals shimmed by the library (`ShimDOMException`,
     //   `ShimDOMStringList`, `ShimEvent`, `ShimCustomEvent`, `ShimEventTarget`)